refactor replication tests
